ledger.checkout: move place_order_submission to utils
