+++ wildlifelicensing/apps/payments/views.py
@@ -11,8 +11,7 @@ from django.utils.http import urlencode
 from django.conf import settings
 from django.utils import timezone
 
-from ledger.checkout.utils import create_basket_session, create_checkout_session
-from ledger.checkout.views import place_order_submission
+from ledger.checkout.utils import create_basket_session, create_checkout_session, place_order_submission
 
 from wildlifelicensing.apps.applications.models import Application
